requirements.txt
# requirements.txt

pandas
requests
networkx
matplotlib
contextily
numPy
python-dotenv
tqdm
pytest
pyproj